Move SE's RichTextBox to its own class
