+++ src/ui/Controls/SETextBox.cs
@@ -1,15 +1,8 @@
 ﻿using Nikse.SubtitleEdit.Core;
 using Nikse.SubtitleEdit.Core.Common;
 using Nikse.SubtitleEdit.Core.Enums;
-using Nikse.SubtitleEdit.Core.Interfaces;
-using Nikse.SubtitleEdit.Core.SpellCheck;
-using Nikse.SubtitleEdit.Logic;
-using Nikse.SubtitleEdit.Logic.SpellCheck;
 using System;
-using System.Collections.Generic;
 using System.Drawing;
-using System.Linq;
-using System.Text.RegularExpressions;
 using System.Threading.Tasks;
 using System.Windows.Forms;
 
@@ -18,7 +11,7 @@ namespace Nikse.SubtitleEdit.Controls
     /// <summary>
     /// TextBox that can be either a normal text box or a rich text box.
     /// </summary>
-    public sealed class SETextBox : Panel, IDoSpell
+    public sealed class SETextBox : Panel
     {
         public new event EventHandler TextChanged;
         public new event KeyEventHandler KeyDown;
@@ -28,47 +21,8 @@ namespace Nikse.SubtitleEdit.Controls
         public new event EventHandler Leave;
         public new event MouseEventHandler MouseMove;
 
-        private bool _checkRtfChange = true;
-        private RichTextBox _uiTextBox;
+        private AdvancedTextBox _uiTextBox;
         private SimpleTextBox _simpleTextBox;
-        private int _mouseMoveSelectionLength;
-
-        private bool _isLiveSpellCheckEnabled => Configuration.Settings.Tools.LiveSpellCheck;
-        private Hunspell _hunspell;
-        private SpellCheckWordLists _spellCheckWordLists;
-        private List<SpellCheckWord> _words;
-        private List<SpellCheckWord> _wrongWords;
-        private List<string> _skipAllList;
-        private HashSet<string> _skipOnceList;
-        private SpellCheckWord _currentWord;
-        private string _currentDictionary;
-        private string _uiTextBoxOldText;
-
-        private static readonly char[] SplitChars = { ' ', '.', ',', '?', '!', ':', ';', '"', '“', '”', '(', ')', '[', ']', '{', '}', '|', '<', '>', '/', '+', '\r', '\n', '\b', '¿', '¡', '…', '—', '–', '♪', '♫', '„', '«', '»', '‹', '›', '؛', '،', '؟' };
-
-        public int CurrentLineIndex { get; set; }
-        public string CurrentLanguage { get; private set; }
-        public bool LanguageChanged { get; set; }
-        public bool IsWrongWord { get; set; }
-        public bool IsSpellCheckerInitialized { get; set; }
-        public bool IsDictionaryDownloaded { get; set; } = true;
-        public bool IsSpellCheckRequested { get; set; }
-
-        public class SuggestionParameter
-        {
-            public string InputWord { get; set; }
-            public List<string> Suggestions { get; set; }
-            public Hunspell Hunspell { get; set; }
-            public bool Success { get; set; }
-
-            public SuggestionParameter(string word, Hunspell hunspell)
-            {
-                InputWord = word;
-                Suggestions = new List<string>();
-                Hunspell = hunspell;
-                Success = false;
-            }
-        }
 
         public SETextBox()
         {
@@ -103,81 +57,9 @@ namespace Nikse.SubtitleEdit.Controls
             _uiTextBox = null;
             if (useSyntaxColoring)
             {
-                _uiTextBox = new RichTextBox { BorderStyle = BorderStyle.None, Multiline = true };
+                _uiTextBox = new AdvancedTextBox{ BorderStyle = BorderStyle.None, Multiline = true };
                 InitializeBackingControl(_uiTextBox);
                 UpdateFontAndColors(_uiTextBox);
-                SetTextPosInRtbIfCentered();
-
-                _uiTextBox.TextChanged += TextChangedHighlight;
-                _uiTextBox.KeyPress += UiTextBox_KeyPress;
-                _uiTextBox.KeyDown += UiTextBox_KeyDown;
-                _uiTextBox.MouseDown += UiTextBox_MouseDown;
-
-                _uiTextBox.HandleCreated += (sender, args) =>
-                {
-                    SetTextPosInRtbIfCentered();
-                };
-                _uiTextBox.MouseDown += (sender, args) =>
-                {
-                    // avoid selection when centered and clicking to the left
-                    var charIndex = _uiTextBox.GetCharIndexFromPosition(args.Location);
-                    if (Configuration.Settings.General.CenterSubtitleInTextBox &&
-                        _mouseMoveSelectionLength == 0 &&
-                        (charIndex == 0 || charIndex >= 0 && _uiTextBox.Text[charIndex - 1] == '\n'))
-                    {
-                        _uiTextBox.SelectionLength = 0;
-                    }
-                };
-                _uiTextBox.MouseMove += (sender, args) =>
-                {
-                    _mouseMoveSelectionLength = _uiTextBox.SelectionLength;
-                };
-                _uiTextBox.KeyDown += (sender, args) =>
-                {
-                    // fix annoying "beeps" when moving cursor position
-                    if ((args.KeyData == Keys.Left || args.KeyData == Keys.PageUp) && _uiTextBox.SelectionStart == 0)
-                    {
-                        args.SuppressKeyPress = true;
-                    }
-                    else if (args.KeyData == Keys.Up && _uiTextBox.SelectionStart <= _uiTextBox.Text.IndexOf('\n'))
-                    {
-                        args.SuppressKeyPress = true;
-                    }
-                    else if (args.KeyData == Keys.Home && (_uiTextBox.SelectionStart == 0 || _uiTextBox.SelectionStart > 0 && _uiTextBox.Text[_uiTextBox.SelectionStart - 1] == '\n'))
-                    {
-                        args.SuppressKeyPress = true;
-                    }
-                    else if (args.KeyData == (Keys.Home | Keys.Control) && _uiTextBox.SelectionStart == 0)
-                    {
-                        args.SuppressKeyPress = true;
-                    }
-                    else if (args.KeyData == Keys.End && (_uiTextBox.SelectionStart >= _uiTextBox.Text.Length || _uiTextBox.SelectionStart + 1 < _uiTextBox.Text.Length && _uiTextBox.Text[_uiTextBox.SelectionStart] == '\n'))
-                    {
-                        args.SuppressKeyPress = true;
-                    }
-                    else if (args.KeyData == (Keys.End | Keys.Control) && _uiTextBox.SelectionStart >= _uiTextBox.Text.Length)
-                    {
-                        args.SuppressKeyPress = true;
-                    }
-                    else if (args.KeyData == Keys.Right && _uiTextBox.SelectionStart >= _uiTextBox.Text.Length)
-                    {
-                        if (_isLiveSpellCheckEnabled)
-                        {
-                            IsSpellCheckRequested = true;
-                            TextChangedHighlight(this, EventArgs.Empty);
-                        }
-
-                        args.SuppressKeyPress = true;
-                    }
-                    else if (args.KeyData == Keys.Down && _uiTextBox.SelectionStart >= _uiTextBox.Text.Length)
-                    {
-                        args.SuppressKeyPress = true;
-                    }
-                    else if (args.KeyData == Keys.PageDown && _uiTextBox.SelectionStart >= _uiTextBox.Text.Length)
-                    {
-                        args.SuppressKeyPress = true;
-                    }
-                };
             }
             else
             {
@@ -235,7 +117,7 @@ namespace Nikse.SubtitleEdit.Controls
             }
         }
 
-        private bool _fixedArabicComma;
+
         public override string Text
         {
             get
@@ -247,14 +129,7 @@ namespace Nikse.SubtitleEdit.Controls
 
                 if (_uiTextBox != null)
                 {
-                    var s = _uiTextBox.Text;
-                    if (_fixedArabicComma)
-                    {
-                        s = s.Replace("\u202A،", "،");
-                        s = s.Replace("،\u202A", "،");
-                    }
-
-                    return string.Join(Environment.NewLine, s.SplitToLines());
+                    return _uiTextBox.Text;
                 }
 
                 return string.Empty;
@@ -269,24 +144,7 @@ namespace Nikse.SubtitleEdit.Controls
 
                 if (_uiTextBox != null)
                 {
-                    _fixedArabicComma = false;
-                    var s = value;
-                    if (!Configuration.Settings.General.RightToLeftMode && !s.ContainsUnicodeControlChars())
-                    {
-                        string textNoTags = HtmlUtil.RemoveHtmlTags(s, true);
-                        if (textNoTags.EndsWith('،'))
-                        {
-                            s = Regex.Replace(s, @"،(?=(?:<[^>]+>)?(?:{[^}]+})?$)", "\u202A،");
-                        }
-                        else if (textNoTags.StartsWith('،'))
-                        {
-                            s = Regex.Replace(s, @"(?<=^(?:{[^}]+})?(?:<[^>]+>)?)،", "،\u202A");
-                        }
-
-                        _fixedArabicComma = true;
-                    }
-
-                    _uiTextBox.Text = string.Join("\n", s.SplitToLines());
+                    _uiTextBox.Text = value;
                 }
             }
         }
@@ -302,20 +160,8 @@ namespace Nikse.SubtitleEdit.Controls
 
                 if (_uiTextBox != null)
                 {
-                    var text = _uiTextBox.Text;
-                    var extra = 0;
-                    var target = _uiTextBox.SelectionStart;
-                    for (int i = 0; i < target && i < text.Length; i++)
-                    {
-                        if (text[i] == '\n')
-                        {
-                            extra++;
-                        }
-                    }
-
-                    return target + extra;
+                    return _uiTextBox.SelectionStart;
                 }
-
 
                 return 0;
             }
@@ -327,17 +173,7 @@ namespace Nikse.SubtitleEdit.Controls
                 }
                 else if (_uiTextBox != null)
                 {
-                    var text = _uiTextBox.Text;
-                    var extra = 0;
-                    for (int i = 0; i < value && i < text.Length; i++)
-                    {
-                        if (text[i] == '\n')
-                        {
-                            extra++;
-                        }
-                    }
-
-                    _uiTextBox.SelectionStart = value - extra;
+                    _uiTextBox.SelectionStart = value;
                 }
             }
         }
@@ -353,24 +189,7 @@ namespace Nikse.SubtitleEdit.Controls
 
                 if (_uiTextBox != null)
                 {
-                    var target = _uiTextBox.SelectionLength;
-                    if (target == 0)
-                    {
-                        return 0;
-                    }
-
-                    var text = _uiTextBox.Text;
-                    var extra = 0;
-                    var start = _uiTextBox.SelectionStart;
-                    for (int i = start; i < target + start && i < text.Length; i++)
-                    {
-                        if (text[i] == '\n')
-                        {
-                            extra++;
-                        }
-                    }
-
-                    return target + extra;
+                    return _uiTextBox.SelectionLength;
                 }
 
                 return 0;
@@ -383,14 +202,7 @@ namespace Nikse.SubtitleEdit.Controls
                 }
                 else if (_uiTextBox != null)
                 {
-                    var target = value;
-                    if (target == 0)
-                    {
-                        _uiTextBox.SelectionLength = 0;
-                        return;
-                    }
-
-                    _uiTextBox.SelectionLength = GetIndexWithLineBreak(value);
+                    _uiTextBox.SelectionLength = value;
                 }
             }
         }
@@ -428,7 +240,12 @@ namespace Nikse.SubtitleEdit.Controls
                     return _simpleTextBox.SelectedText;
                 }
 
-                return _uiTextBox != null ? string.Join(Environment.NewLine, _uiTextBox.SelectedText.SplitToLines()) : string.Empty;
+                if (_uiTextBox != null)
+                {
+                    return _uiTextBox.SelectedText;
+                }
+
+                return string.Empty;
             }
             set
             {
@@ -559,21 +376,6 @@ namespace Nikse.SubtitleEdit.Controls
             }
         }
 
-        private int GetIndexWithLineBreak(int index)
-        {
-            var text = _uiTextBox.Text;
-            var extra = 0;
-            for (int i = 0; i < index && i < text.Length; i++)
-            {
-                if (text[i] == '\n')
-                {
-                    extra++;
-                }
-            }
-
-            return index - extra;
-        }
-
         public int GetCharIndexFromPosition(Point pt)
         {
             if (_simpleTextBox != null)
@@ -693,706 +495,206 @@ namespace Nikse.SubtitleEdit.Controls
             }
         }
 
-        private void DoFormattingActionOnRtb(Action formattingAction)
-        {
-            _uiTextBox.BeginRichTextBoxUpdate();
-            var start = _uiTextBox.SelectionStart;
-            var length = _uiTextBox.SelectionLength;
-            formattingAction();
-            _uiTextBox.SelectionStart = start;
-            _uiTextBox.SelectionLength = length;
-            _uiTextBox.EndRichTextBoxUpdate();
-        }
-
-        private void SetTextPosInRtbIfCentered()
-        {
-            if (Configuration.Settings.General.CenterSubtitleInTextBox)
-            {
-                DoFormattingActionOnRtb(() =>
-                {
-                    _uiTextBox.SelectAll();
-                    _uiTextBox.SelectionAlignment = HorizontalAlignment.Center;
-                });
-            }
-        }
-
-        private bool ContainsTags() =>
-            _uiTextBox.Text.Length != HtmlUtil.RemoveHtmlTags(_uiTextBox.Text, true).Length;
-
-        private void TagsChangedCheck()
-        {
-            if (!string.IsNullOrEmpty(_uiTextBoxOldText)
-                && HtmlUtil.RemoveHtmlTags(_uiTextBoxOldText, true) == HtmlUtil.RemoveHtmlTags(_uiTextBox.Text, true))
-            {
-                IsSpellCheckRequested = true;
-                _uiTextBoxOldText = _uiTextBox.Text;
-            }
-        }
-
-        private void TextChangedHighlight(object sender, EventArgs e)
-        {
-            if (_checkRtfChange)
-            {
-                _checkRtfChange = false;
-
-                // Request spell check if there is a change in tags to update highlighting indices.
-                if (_isLiveSpellCheckEnabled)
-                {
-                    TagsChangedCheck();
-                }
-
-                if (ContainsTags() || IsSpellCheckRequested)
-                {
-                    DoFormattingActionOnRtb(HighlightHtmlText);
-                }
-
-                _checkRtfChange = true;
-            }
-        }
-
-        private void HighlightHtmlText()
-        {
-            _uiTextBox.SelectAll();
-            _uiTextBox.SelectionColor = _uiTextBox.ForeColor;
-            _uiTextBox.SelectionBackColor = _uiTextBox.BackColor;
-
-            bool htmlTagOn = false;
-            bool htmlTagFontOn = false;
-            int htmlTagStart = -1;
-            bool assaTagOn = false;
-            bool assaPrimaryColorTagOn = false;
-            bool assaSecondaryColorTagOn = false;
-            bool assaBorderColorTagOn = false;
-            bool assaShadowColorTagOn = false;
-            var assaTagStart = -1;
-            int tagOn = -1;
-            var text = _uiTextBox.Text;
-            var textLength = text.Length;
-            int i = 0;
-
-            while (i < textLength)
-            {
-                var ch = text[i];
-                if (assaTagOn)
-                {
-                    if (ch == '}' && tagOn >= 0)
-                    {
-                        assaTagOn = false;
-                        _uiTextBox.SelectionStart = assaTagStart;
-                        _uiTextBox.SelectionLength = i - assaTagStart + 1;
-                        _uiTextBox.SelectionColor = Configuration.Settings.General.SubtitleTextBoxAssColor;
-                        if (assaTagStart >= 0)
-                        {
-                            if (assaPrimaryColorTagOn)
-                            {
-                                var colorTag = text.IndexOf("\\c", assaTagStart, StringComparison.OrdinalIgnoreCase) != -1 ? "\\c" : "\\1c";
-                                SetAssaColor(text, assaTagStart, colorTag);
-                                assaPrimaryColorTagOn = false;
-                            }
-
-                            if (assaSecondaryColorTagOn)
-                            {
-                                SetAssaColor(text, assaTagStart, "\\2c");
-                                assaSecondaryColorTagOn = false;
-                            }
-
-                            if (assaBorderColorTagOn)
-                            {
-                                SetAssaColor(text, assaTagStart, "\\3c");
-                                assaBorderColorTagOn = false;
-                            }
-
-                            if (assaShadowColorTagOn)
-                            {
-                                SetAssaColor(text, assaTagStart, "\\4c");
-                                assaShadowColorTagOn = false;
-                            }
-                        }
-
-                        assaTagStart = -1;
-                    }
-                }
-                else if (htmlTagOn)
-                {
-                    if (ch == '>' && tagOn >= 0)
-                    {
-                        htmlTagOn = false;
-                        _uiTextBox.SelectionStart = htmlTagStart;
-                        _uiTextBox.SelectionLength = i - htmlTagStart + 1;
-                        _uiTextBox.SelectionColor = Configuration.Settings.General.SubtitleTextBoxHtmlColor;
-                        if (htmlTagFontOn && htmlTagStart >= 0)
-                        {
-                            SetHtmlColor(text, htmlTagStart);
-                            htmlTagFontOn = false;
-                        }
-                        htmlTagStart = -1;
-                    }
-                }
-                else if (ch == '{' && i < textLength - 1 && text[i + 1] == '\\' && text.IndexOf('}', i) > 0)
-                {
-                    var s = text.Substring(i);
-                    assaTagOn = true;
-                    tagOn = i;
-                    assaTagStart = i;
-                    assaPrimaryColorTagOn = s.Contains("\\c", StringComparison.OrdinalIgnoreCase) || s.Contains("\\1c", StringComparison.OrdinalIgnoreCase);
-                    assaSecondaryColorTagOn = s.Contains("\\2c", StringComparison.OrdinalIgnoreCase);
-                    assaBorderColorTagOn = s.Contains("\\3c", StringComparison.OrdinalIgnoreCase);
-                    assaShadowColorTagOn = s.Contains("\\4c", StringComparison.OrdinalIgnoreCase);
-                }
-                else if (ch == '<')
-                {
-                    var s = text.Substring(i);
-                    if (s.StartsWith("<i>", StringComparison.OrdinalIgnoreCase) ||
-                        s.StartsWith("<b>", StringComparison.OrdinalIgnoreCase) ||
-                        s.StartsWith("<u>", StringComparison.OrdinalIgnoreCase) ||
-                        s.StartsWith("</i>", StringComparison.OrdinalIgnoreCase) ||
-                        s.StartsWith("</b>", StringComparison.OrdinalIgnoreCase) ||
-                        s.StartsWith("</u>", StringComparison.OrdinalIgnoreCase) ||
-                        s.StartsWith("<box>", StringComparison.OrdinalIgnoreCase) ||
-                        s.StartsWith("</box>", StringComparison.OrdinalIgnoreCase) ||
-                        s.StartsWith("</font>", StringComparison.OrdinalIgnoreCase) ||
-                        (s.StartsWith("<font ", StringComparison.OrdinalIgnoreCase) &&
-                         text.IndexOf("</font>", i, StringComparison.OrdinalIgnoreCase) > 0))
-                    {
-                        htmlTagOn = true;
-                        htmlTagStart = i;
-                        htmlTagFontOn = s.StartsWith("<font ", StringComparison.OrdinalIgnoreCase);
-                        tagOn = i;
-                    }
-                }
-
-                i++;
-            }
-
-            if (IsSpellCheckRequested)
-            {
-                DoLiveSpellCheck();
-                if (_wrongWords?.Count > 0)
-                {
-                    foreach (var wrongWord in _wrongWords)
-                    {
-                        _uiTextBox.Select(GetIndexWithLineBreak(wrongWord.Index), wrongWord.Length);
-                        _uiTextBox.SelectionColor = Configuration.Settings.Tools.ListViewSyntaxErrorColor;
-                    }
-                }
-
-                IsSpellCheckRequested = false;
-            }
-        }
-
-        private void SetHtmlColor(string text, int htmlTagStart)
-        {
-            int colorStart = text.IndexOf(" color=", htmlTagStart, StringComparison.OrdinalIgnoreCase);
-            if (colorStart > 0)
-            {
-                colorStart += " color=".Length;
-                if (text[colorStart] == '"' || text[colorStart] == '\'')
-                {
-                    colorStart++;
-                }
-
-                int colorEnd = text.IndexOf('"', colorStart + 1);
-                if (colorEnd > 0)
-                {
-                    var color = text.Substring(colorStart, colorEnd - colorStart);
-                    try
-                    {
-                        Color c;
-                        if (color.StartsWith("rgb(", StringComparison.Ordinal))
-                        {
-                            var arr = color.Remove(0, 4).TrimEnd(')').Split(new[] { ',' }, StringSplitOptions.RemoveEmptyEntries);
-                            c = Color.FromArgb(int.Parse(arr[0]), int.Parse(arr[1]), int.Parse(arr[2]));
-                        }
-                        else
-                        {
-                            c = ColorTranslator.FromHtml(color);
-                        }
-
-                        SetForeColorAndChangeBackColorIfClose(colorStart, colorEnd, c);
-                    }
-                    catch
-                    {
-                        // ignored
-                    }
-                }
-            }
-        }
-
-        private void SetAssaColor(string text, int assaTagStart, string colorTag)
-        {
-            int colorStart = text.IndexOf(colorTag, assaTagStart, StringComparison.OrdinalIgnoreCase);
-            if (colorStart > 0)
-            {
-                colorStart += colorTag.Length;
-                if (text[colorStart] == '&')
-                {
-                    colorStart++;
-                }
-                if (text[colorStart] == 'H')
-                {
-                    colorStart++;
-                }
-
-                int colorEnd = text.IndexOf('&', colorStart + 1);
-                if (colorEnd > 0)
-                {
-                    var color = text.Substring(colorStart, colorEnd - colorStart);
-                    try
-                    {
-                        if (color.Length == 6)
-                        {
-                            var rgbColor = string.Concat("#", color[4], color[5], color[2], color[3], color[0], color[1]); var c = ColorTranslator.FromHtml(rgbColor);
-                            SetForeColorAndChangeBackColorIfClose(colorStart, colorEnd, c);
-                        }
-                    }
-                    catch
-                    {
-                        // ignored
-                    }
-                }
-            }
-        }
-
-        private void SetForeColorAndChangeBackColorIfClose(int colorStart, int colorEnd, Color c)
-        {
-            var backColor = _uiTextBox.BackColor;
-            _uiTextBox.SelectionStart = colorStart;
-            _uiTextBox.SelectionLength = colorEnd - colorStart;
-            _uiTextBox.SelectionColor = c;
-
-            var diff = Math.Abs(c.R - backColor.R) + Math.Abs(c.G - backColor.G) + Math.Abs(c.B - backColor.B);
-            if (diff < 60)
-            {
-                _uiTextBox.SelectionBackColor = Color.FromArgb(byte.MaxValue - c.R, byte.MaxValue - c.G, byte.MaxValue - c.B, byte.MaxValue - c.R);
-            }
-        }
-
         #region LiveSpellCheck
+
+        public int CurrentLineIndex
+        {
+            get
+            {
+                if (_uiTextBox != null)
+                {
+                    return _uiTextBox.CurrentLineIndex;
+                }
+
+                return 0;
+            }
+            set
+            {
+                if (_uiTextBox != null)
+                {
+                    _uiTextBox.CurrentLineIndex = value;
+                }
+            }
+        }
+
+        public string CurrentLanguage
+        {
+            get
+            {
+                if (_uiTextBox != null)
+                {
+                    return _uiTextBox.CurrentLanguage;
+                }
+
+                return string.Empty;
+            }
+            set
+            {
+                if (_uiTextBox != null)
+                {
+                    _uiTextBox.CurrentLanguage = value;
+                }
+            }
+        }
+
+        public bool LanguageChanged
+        {
+            get
+            {
+                if (_uiTextBox != null)
+                {
+                    return _uiTextBox.LanguageChanged;
+                }
+
+                return false;
+            }
+            set
+            {
+                if (_uiTextBox != null)
+                {
+                    _uiTextBox.LanguageChanged = value;
+                }
+            }
+        }
+
+        public bool IsWrongWord
+        {
+            get
+            {
+                if (_uiTextBox != null)
+                {
+                    return _uiTextBox.IsWrongWord;
+                }
+
+                return false;
+            }
+            set
+            {
+                if (_uiTextBox != null)
+                {
+                    _uiTextBox.IsWrongWord = value;
+                }
+            }
+        }
+
+        public bool IsSpellCheckerInitialized
+        {
+            get
+            {
+                if (_uiTextBox != null)
+                {
+                    return _uiTextBox.IsSpellCheckerInitialized;
+                }
+
+                return false;
+            }
+            set
+            {
+                if (_uiTextBox != null)
+                {
+                    _uiTextBox.IsSpellCheckerInitialized = value;
+                }
+            }
+        }
+
+        public bool IsDictionaryDownloaded
+        {
+            get
+            {
+                if (_uiTextBox != null)
+                {
+                    return _uiTextBox.IsDictionaryDownloaded;
+                }
+
+                return true;
+            }
+            set
+            {
+                if (_uiTextBox != null)
+                {
+                    _uiTextBox.IsDictionaryDownloaded = value;
+                }
+            }
+        }
+
+        public bool IsSpellCheckRequested
+        {
+            get
+            {
+                if (_uiTextBox != null)
+                {
+                    return _uiTextBox.IsSpellCheckRequested;
+                }
+
+                return false;
+            }
+            set
+            {
+                if (_uiTextBox != null)
+                {
+                    _uiTextBox.IsSpellCheckRequested = value;
+                }
+            }
+        }
 
         public async Task CheckForLanguageChange(Subtitle subtitle)
         {
-            var detectedLanguage = LanguageAutoDetect.AutoDetectGoogleLanguage(subtitle);
-            if (CurrentLanguage != detectedLanguage)
-            {
-                DisposeHunspellAndDictionaries();
-                await InitializeLiveSpellCheck(subtitle, CurrentLineIndex);
-            }
-        }
-
-        public async Task InitializeLiveSpellCheck(Subtitle subtitle, int lineNumber)
-        {
-            if (_uiTextBox is null || lineNumber < 0)
+            if (_uiTextBox is null)
             {
                 return;
             }
 
-            if (_spellCheckWordLists is null && _hunspell is null)
+            await _uiTextBox.CheckForLanguageChange(subtitle);
+        }
+
+        public async Task InitializeLiveSpellCheck(Subtitle subtitle, int lineNumber)
+        {
+            if (_uiTextBox is null)
             {
-                var detectedLanguage = LanguageAutoDetect.AutoDetectGoogleLanguage(subtitle);
-                var downloadedDictionaries = Utilities.GetDictionaryLanguagesCultureNeutral();
-                var isDictionaryAvailable = false;
-                foreach (var downloadedDictionary in downloadedDictionaries)
-                {
-                    if (downloadedDictionary.Contains($"[{detectedLanguage}]"))
-                    {
-                        isDictionaryAvailable = true;
-                        break;
-                    }
-                }
-
-                if (isDictionaryAvailable)
-                {
-                    IsDictionaryDownloaded = true;
-
-                    var languageName = LanguageAutoDetect.AutoDetectLanguageName(string.Empty, subtitle);
-                    if (languageName.Split(new char[] { '_', '-' })[0] != detectedLanguage)
-                    {
-                        return;
-                    }
-
-                    await LoadDictionariesAsync(languageName);
-                    IsSpellCheckerInitialized = true;
-                    IsSpellCheckRequested = true;
-                    TextChangedHighlight(this, EventArgs.Empty);
-                }
-                else
-                {
-                    IsDictionaryDownloaded = false;
-                }
-
-                LanguageChanged = true;
-                CurrentLanguage = detectedLanguage;
+                return;
             }
-        }
 
-        private async Task LoadDictionariesAsync(string languageName)
-        {
-            await Task.Run(() => LoadDictionaries(languageName));
-        }
-
-        private void LoadDictionaries(string languageName)
-        {
-            var dictionaryFolder = Utilities.DictionaryFolder;
-            string dictionary = Utilities.DictionaryFolder + languageName;
-            _spellCheckWordLists = new SpellCheckWordLists(dictionaryFolder, languageName, this);
-            _skipAllList = new List<string>();
-            _skipOnceList = new HashSet<string>();
-            LoadHunspell(dictionary);
-        }
-
-        private void LoadHunspell(string dictionary)
-        {
-            _currentDictionary = dictionary;
-            _hunspell?.Dispose();
-            _hunspell = null;
-            _hunspell = Hunspell.GetHunspell(dictionary);
+            await _uiTextBox.InitializeLiveSpellCheck(subtitle, lineNumber);
         }
 
         public void DisposeHunspellAndDictionaries()
         {
-            if (IsSpellCheckerInitialized)
+            if (_uiTextBox is null)
             {
-                _skipAllList = null;
-                _skipOnceList = null;
-                _spellCheckWordLists = null;
-                _words = null;
-                _wrongWords = null;
-                _currentWord = null;
-                _currentDictionary = null;
-                CurrentLanguage = null;
-                _hunspell?.Dispose();
-                _hunspell = null;
-                IsWrongWord = false;
-                IsSpellCheckerInitialized = false;
-
-                if (Configuration.Settings.General.SubtitleTextBoxSyntaxColor)
-                {
-                    IsSpellCheckRequested = true;
-                    TextChangedHighlight(this, EventArgs.Empty);
-                }
+                return;
             }
+
+            _uiTextBox.DisposeHunspellAndDictionaries();
         }
 
         public void DoLiveSpellCheck()
         {
-            if (IsSpellCheckerInitialized && IsSpellCheckRequested)
+            if (_uiTextBox is null)
             {
-                _words = GetWords(Text);
-                SetWrongWords();
-            }
-        }
-
-        private List<SpellCheckWord> GetWords(string s)
-        {
-            s = _spellCheckWordLists.ReplaceHtmlTagsWithBlanks(s);
-            s = _spellCheckWordLists.ReplaceAssTagsWithBlanks(s);
-            s = _spellCheckWordLists.ReplaceKnownWordsOrNamesWithBlanks(s);
-            return SpellCheckWordLists.Split(s);
-        }
-
-        private void SetWrongWords()
-        {
-            _wrongWords = new List<SpellCheckWord>();
-
-            for (int i = 0; i < _words.Count; i++)
-            {
-                var currentWord = _words[i];
-                var currentWordText = _words[i].Text;
-                int minLength = 2;
-                if (Configuration.Settings.Tools.CheckOneLetterWords)
-                {
-                    minLength = 1;
-                }
-
-                string key = CurrentLineIndex + "-" + currentWord.Text + "-" + currentWord.Index;
-                if (DoSpell(currentWordText) || Utilities.IsNumber(currentWordText) || _skipAllList.Contains(currentWordText)
-                    || _skipOnceList.Contains(key) || _spellCheckWordLists.HasUserWord(currentWordText) || _spellCheckWordLists.HasName(currentWordText)
-                    || currentWordText.Length < minLength || currentWordText == "&")
-                {
-                    continue;
-                }
-
-                string prefix = string.Empty;
-                string postfix = string.Empty;
-                if (currentWordText.RemoveControlCharacters().Trim().Length >= minLength)
-                {
-                    if (currentWordText.Length > 0)
-                    {
-                        var trimChars = "'`*#\u200E\u200F\u202A\u202B\u202C\u202D\u202E\u200B\uFEFF";
-                        var charHit = true;
-                        while (charHit)
-                        {
-                            charHit = false;
-                            foreach (char c in trimChars)
-                            {
-                                if (currentWordText.StartsWith(c))
-                                {
-                                    prefix += c;
-                                    currentWordText = currentWordText.Substring(1);
-                                    charHit = true;
-                                }
-                                if (currentWordText.EndsWith(c))
-                                {
-                                    postfix = c + postfix;
-                                    currentWordText = currentWordText.Remove(currentWordText.Length - 1);
-                                    charHit = true;
-                                }
-                            }
-                        }
-                    }
-                }
-
-                if (prefix == "'" && currentWordText.Length >= 1 && (DoSpell(prefix + currentWordText) || _spellCheckWordLists.HasUserWord(prefix + currentWordText)))
-                {
-                    continue;
-                }
-                else if (currentWordText.Length > 1)
-                {
-                    if ("`'".Contains(currentWordText[currentWordText.Length - 1]) && DoSpell(currentWordText.TrimEnd('\'').TrimEnd('`')))
-                    {
-                        continue;
-                    }
-
-                    if (currentWordText.EndsWith("'s", StringComparison.Ordinal) && currentWordText.Length > 4
-                    && DoSpell(currentWordText.TrimEnd('s').TrimEnd('\'')))
-                    {
-                        continue;
-                    }
-
-                    if (currentWordText.EndsWith('\'') && DoSpell(currentWordText.TrimEnd('\'')))
-                    {
-                        continue;
-                    }
-
-                    string removeUnicode = currentWordText.Replace("\u200b", string.Empty); // zero width space
-                    removeUnicode = removeUnicode.Replace("\u2060", string.Empty); // word joiner
-                    removeUnicode = removeUnicode.Replace("\ufeff", string.Empty); // zero width no-break space
-                    if (DoSpell(removeUnicode))
-                    {
-                        continue;
-                    }
-
-                    if (i > 0 && i < _words.Count && _words.ElementAtOrDefault(i + 1) != null &&
-                        _words[i - 1].Text.ToLowerInvariant() == "www" &&
-                        (_words[i + 1].Text.ToLowerInvariant() == "com" ||
-                         _words[i + 1].Text.ToLowerInvariant() == "org" ||
-                         _words[i + 1].Text.ToLowerInvariant() == "net") &&
-                        Text.IndexOf(_words[i - 1].Text + "." + currentWordText + "." + 
-                                     _words[i + 1].Text, StringComparison.OrdinalIgnoreCase) >= 0)
-                    {
-                        continue; // do not spell check urls
-                    }
-
-                    if (CurrentLanguage == "ar")
-                    {
-                        var trimmed = currentWordText.Trim('0', '1', '2', '3', '4', '5', '6', '7', '8', '9', '.', ',', '،');
-                        if (trimmed != currentWordText)
-                        {
-                            if (_spellCheckWordLists.HasName(trimmed) || _skipAllList.Contains(trimmed.ToUpperInvariant())
-                                || _spellCheckWordLists.HasUserWord(trimmed) || DoSpell(trimmed))
-                            {
-                                continue;
-                            }
-                        }
-                    }
-
-                    // check if dash concatenated word with previous or next word is in spell check dictionary
-                    if (i > 0 && (Text[currentWord.Index - 1] == '-' || Text[currentWord.Index - 1] == '‑'))
-                    {
-                        var wordWithDash = _words[i - 1].Text + "-" + currentWordText;
-                        if (DoSpell(wordWithDash))
-                        {
-                            continue;
-                        }
-
-                        wordWithDash = _words[i - 1].Text + "‑" + currentWordText; // non break hyphen
-                        if (DoSpell(wordWithDash) || _spellCheckWordLists.HasUserWord(wordWithDash)
-                        || _spellCheckWordLists.HasUserWord(wordWithDash.Replace("‑", "-")) || _spellCheckWordLists.HasUserWord("-" + currentWordText))
-                        {
-                            continue;
-                        }
-
-                    }
-
-                    if (i < _words.Count - 1 && _words[i + 1].Index - 1 < Text.Length &&
-                        (Text[_words[i + 1].Index - 1] == '-' || Text[_words[i + 1].Index - 1] == '‑'))
-                    {
-                        var wordWithDash = currentWordText + "-" + _words[i + 1].Text;
-                        if (DoSpell(wordWithDash))
-                        {
-                            continue;
-                        }
-
-                        wordWithDash = currentWordText + "‑" + _words[i + 1].Text; // non break hyphen
-                        if (DoSpell(wordWithDash) || _spellCheckWordLists.HasUserWord(wordWithDash) || _spellCheckWordLists.HasUserWord(wordWithDash.Replace("‑", "-")))
-                        {
-                            continue;
-                        }
-                    }
-                }
-                else
-                {
-                    if (currentWordText == "'")
-                    {
-                        continue;
-                    }
-                    else if (CurrentLanguage == "en " && (currentWordText.Equals("a", StringComparison.OrdinalIgnoreCase) || currentWordText == "I"))
-                    {
-                        continue;
-                    }
-                    else if (CurrentLanguage == "da" && currentWordText.Equals("i", StringComparison.OrdinalIgnoreCase))
-                    {
-                        continue;
-                    }
-                }
-
-                if (Configuration.Settings.Tools.SpellCheckEnglishAllowInQuoteAsIng && CurrentLanguage == "en"
-                    && _words[i].Text.EndsWith("in'", StringComparison.OrdinalIgnoreCase) && DoSpell(currentWordText.TrimEnd('\'') + "g"))
-                {
-                    continue;
-                }
-
-                _wrongWords.Add(currentWord);
-            }
-        }
-
-        public bool DoSpell(string word)
-        {
-            return _hunspell.Spell(word);
-        }
-
-        private void UiTextBox_KeyDown(object sender, KeyEventArgs e)
-        {
-            if (_isLiveSpellCheckEnabled && e.KeyCode == Keys.Apps && _wrongWords?.Count > 0)
-            {
-                var cursorPos = _uiTextBox.SelectionStart;
-                var wrongWord = _wrongWords.Where(word => cursorPos > GetIndexWithLineBreak(word.Index) && cursorPos < GetIndexWithLineBreak(word.Index) + word.Length).FirstOrDefault();
-                if (wrongWord != null)
-                {
-                    IsWrongWord = true;
-                    _currentWord = wrongWord;
-                }
-                else
-                {
-                    IsWrongWord = false;
-                }
-            }
-        }
-
-        private void UiTextBox_KeyPress(object sender, KeyPressEventArgs e)
-        {
-            if (_isLiveSpellCheckEnabled && (SplitChars.Contains(e.KeyChar)
-                && _uiTextBox.SelectionStart == _uiTextBox.Text.Length || _uiTextBox.SelectionStart != _uiTextBox.Text.Length))
-            {
-                IsSpellCheckRequested = true;
-                if (e.KeyChar == '\b' || e.KeyChar == '\r' || e.KeyChar == '\n')
-                {
-                    TextChangedHighlight(this, EventArgs.Empty);
-                }
-            }
-        }
-
-        private void UiTextBox_MouseDown(object sender, MouseEventArgs e)
-        {
-            if (_wrongWords?.Count > 0 && e.Clicks == 1 && e.Button == MouseButtons.Right)
-            {
-                int positionToSearch = _uiTextBox.GetCharIndexFromPosition(new Point(e.X, e.Y));
-                var wrongWord = _wrongWords.Where(word => positionToSearch > GetIndexWithLineBreak(word.Index) && positionToSearch < GetIndexWithLineBreak(word.Index) + word.Length).FirstOrDefault();
-                if (wrongWord != null)
-                {
-                    IsWrongWord = true;
-                    _currentWord = wrongWord;
-                }
-                else
-                {
-                    IsWrongWord = false;
-                }
-            }
-        }
-
-        private List<string> DoSuggest(string word)
-        {
-            var parameter = new SuggestionParameter(word, _hunspell);
-            var suggestThread = new System.Threading.Thread(DoWork);
-            suggestThread.Start(parameter);
-            suggestThread.Join(3000); // wait max 3 seconds
-            suggestThread.Abort();
-            if (!parameter.Success)
-            {
-                LoadHunspell(_currentDictionary);
+                return;
             }
 
-            return parameter.Suggestions;
+            _uiTextBox.DoLiveSpellCheck();
         }
 
         public void AddSuggestionsToMenu()
         {
-            if (_currentWord != null)
+            if (_uiTextBox is null)
             {
-                var suggestions = DoSuggest(_currentWord.Text);
-                if (suggestions?.Count > 0)
-                {
-                    foreach (var suggestion in suggestions)
-                    {
-                        _uiTextBox.ContextMenuStrip.Items.Add(suggestion, null, SuggestionSelected);
-                    }
-                }
+                return;
             }
-        }
 
-        private static void DoWork(object data)
-        {
-            var parameter = (SuggestionParameter)data;
-            parameter.Suggestions = parameter.Hunspell.Suggest(parameter.InputWord);
-            parameter.Success = true;
-        }
-
-        private void SuggestionSelected(object sender, EventArgs e)
-        {
-            IsWrongWord = false;
-            _wrongWords.Remove(_currentWord);
-            var item = (ToolStripItem)sender;
-            var correctWord = item.Text;
-            var text = _uiTextBox.Text;
-            var cursorPos = _uiTextBox.SelectionStart;
-            var wordIndex = GetIndexWithLineBreak(_currentWord.Index);
-            text = text.Remove(wordIndex, _currentWord.Length);
-            text = text.Insert(wordIndex, correctWord);
-            _uiTextBox.Text = text;
-            _uiTextBox.SelectionStart = cursorPos;
-            IsSpellCheckRequested = true;
-            TextChangedHighlight(this, EventArgs.Empty);
+            _uiTextBox.AddSuggestionsToMenu();
         }
 
         public void DoAction(SpellCheckAction action)
         {
-            if (_currentWord != null)
+            if (_uiTextBox is null)
             {
-                switch (action)
-                {
-                    case SpellCheckAction.Skip:
-                        string key = CurrentLineIndex + "-" + _currentWord.Text + "-" + _currentWord.Index;
-                        _skipOnceList.Add(key);
-                        break;
-                    case SpellCheckAction.SkipAll:
-                        _skipAllList.Add(_currentWord.Text);
-                        break;
-                    case SpellCheckAction.AddToDictionary:
-                        _spellCheckWordLists.AddUserWord(_currentWord.Text);
-                        break;
-                    case SpellCheckAction.AddToNames:
-                        _spellCheckWordLists.AddName(_currentWord.Text);
-                        break;
-                }
-
-                if (_wrongWords.Contains(_currentWord))
-                {
-                    IsWrongWord = false;
-                    _wrongWords.Remove(_currentWord);
-                    IsSpellCheckRequested = true;
-                    TextChangedHighlight(this, EventArgs.Empty);
-                }
+                return;
             }
+
+            _uiTextBox.DoAction(action);
         }
 
         #endregion
